+++ tasks/build.js
@@ -1,52 +1,18 @@
 var path = require('path'),
-  fs = require('fs');
+  fs = require('fs'),
+  tools = require('../build/statictools');
 
 module.exports = function(grunt){
 
-
   grunt.registerTask('build', 'Build dist from bower components' , function(){
     var done = this.async();
-    grunt.log.writeln('Fetching files from bower_components...');
-    try {
-      buildGruntConfiguration(grunt, 'bower_components', function(err, configs){
-        if (err) grunt.log.write(JSON.stringify(err));
-        grunt.log.writeln('Loading Skin:', grunt.option('skin')||'default', ' ');
-        loadSkin(grunt, grunt.option('skin'), configs.stylus);
-        grunt.config.set('stylus', configs.stylus );
-        grunt.config.set('uglify', configs.uglify );
-        grunt.file.copy('bower_components/x-tag-core/dist/x-tag-core.min.js','dist/x-tag-core.min.js');
-        grunt.file.copy('build/readme.txt','dist/readme.txt');
-        grunt.file.copy('build/OpenSans-SemiBold.ttf','dist/OpenSans-SemiBold.ttf');
-        grunt.task.run('stylus','uglify');
-        grunt.file.copy('bower_components/x-tag-core/dist/x-tag-core.min.js','dist/x-tag-core.min.js');
-        grunt.file.copy('bower_components/x-tag-core/dist/x-tag-core.js','dist/x-tag-core.js');
-        grunt.file.copy('build/readme.txt','dist/readme.txt');
-        grunt.file.copy('build/OpenSans-SemiBold.ttf','dist/OpenSans-SemiBold.ttf');
-        done();
-      });
-    } catch (e) {
-      grunt.log.error('something has gone terribly wrong.');
-      grunt.log.error(JSON.stringify(e));
-      throw e;
-    }
-  });
 
-  grunt.registerTask('build-dev', 'Build dist from dev repositories' , function(){
-    var done = this.async();
-    grunt.log.writeln('Fetching files from dev-repos...');
     try {
-      buildGruntConfiguration(grunt, 'dev-repos', function(err, configs){
-        if (err) grunt.log.write(err);
-        grunt.log.writeln('Loading Skin:', grunt.option('skin')||'default', ' ');
-        loadSkin(grunt, grunt.option('skin'), configs.stylus);
-        grunt.config.set('stylus', configs.stylus );
-        grunt.config.set('uglify', configs.uglify );
-        grunt.task.run('stylus','uglify');
-        grunt.file.copy('dev-repos/x-tag-core/dist/x-tag-core.min.js','dist/x-tag-core.min.js');
-        grunt.file.copy('dev-repos/x-tag-core/dist/x-tag-core.js','dist/x-tag-core.js');
-        grunt.file.copy('build/readme.txt','dist/readme.txt');
-        grunt.file.copy('build/OpenSans-SemiBold.ttf','dist/OpenSans-SemiBold.ttf');
-        done();
+      var sourcePath = ~process.argv.indexOf('--dev') ? 'dev-repos' : 'bower_components';
+      grunt.log.writeln('Fetching files from ' + sourcePath);
+      buildGruntConfiguration(grunt, sourcePath, function(err, configs){
+        if (err) grunt.log.write(JSON.stringify(err));
+        execGrunt(grunt, sourcePath, configs, done);
       });
     } catch (e) {
       grunt.log.error('something has gone terribly wrong.');
@@ -78,6 +44,19 @@ module.exports = function(grunt){
     grunt.task.run('compress');
   });
 
+}
+
+function execGrunt(grunt, sourcePath, configs, done){
+  grunt.log.writeln('Loading Skin:', grunt.option('skin')||'default', ' ');
+  loadSkin(grunt, grunt.option('skin'), configs.stylus);
+  grunt.config.set('stylus', configs.stylus );
+  grunt.config.set('uglify', configs.uglify );
+  grunt.task.run('stylus','uglify');
+  grunt.file.copy(sourcePath + '/x-tag-core/dist/x-tag-core.min.js','dist/x-tag-core.min.js');
+  grunt.file.copy(sourcePath + '/x-tag-core/dist/x-tag-core.js','dist/x-tag-core.js');
+  grunt.file.copy('build/readme.txt','dist/readme.txt');
+  grunt.file.copy('build/OpenSans-SemiBold.ttf','dist/OpenSans-SemiBold.ttf');
+  done();
 }
 
 function loadSkin(grunt, skinFolder, config){
@@ -137,7 +116,7 @@ function buildGruntConfiguration(grunt, source, callback){
     uglifyConfig = {};
 
   stylusConfig[path.join('dist','brick.css')] = [];
-  uglifyConfig[path.join('dist','brick.js')] = [path.join(source,'x-tag-core','dist','x-tag-core.js')];
+  uglifyConfig[path.join('dist','brick.js')] = [];
 
   grunt.log.debug('spawning bower tasks');
 
@@ -153,48 +132,51 @@ function buildGruntConfiguration(grunt, source, callback){
       grunt.log.debug(result.stdout);
       throw e;
     }
-    var dependencies = bower_data.dependencies;
-    var componentsJson = grunt.file.readJSON('./build/components.json');
-    var components = [];
-    Object.keys(componentsJson).forEach(function(key){
-      components.push(componentsJson[key]);
-    });
 
-    var dKeys = Object.keys(dependencies).filter(function(c){
-      for (var i = 0; i < components.length; i++){
-        if (c.indexOf(components[i])>-1){
-          return true;
-        }
-      };
-    });
+    var dependencies = tools.flattenBowerDependencies(bower_data);
 
-    grunt.log.debug('iterating over component deps');
+    grunt.log.debug('iterating over component dependencies');
 
-    dKeys.forEach(function(k){
+    dependencies.forEach(function(item){
+      k = Object.keys(item)[0];
 
       grunt.log.debug('dependency ' + k);
 
-      var stylusFile = dependencies[k].pkgMeta.main.filter(function(f){
+      var stylusFile = item[k].pkgMeta.main.filter(function(f){
         if(f.indexOf('.styl')>-1){
           return true;
         }
-      })[0];
+      });
 
-      var jsFile = dependencies[k].pkgMeta.main.filter(function(f){
+      var jsFile = item[k].pkgMeta.main.filter(function(f){
         if(f.indexOf('.js')>-1){
           return true;
         }
-      })[0];
+      });
 
       var dest = path.join('dist', k);
 
-      stylusFile = path.join(source, k, stylusFile);
-      stylusConfig[dest + '.css'] = stylusFile
-      stylusConfig[path.join('dist','brick.css')].push(stylusFile);
+      stylusFile.forEach(function(file){
+        file = path.join(source, k, file);
+        if (stylusConfig[dest + '.css']){
+          stylusConfig[dest + '.css'].push(file);
+        }
+        else {
+          stylusConfig[dest + '.css'] = [file];
+        }
+        stylusConfig[path.join('dist','brick.css')].push(file);
+      });
 
-      jsFile = path.join(source, k, jsFile);
-      uglifyConfig[dest + '.js'] = jsFile;
-      uglifyConfig[path.join('dist','brick.js')].push(jsFile);
+      jsFile.forEach(function(file){
+        file = path.join(source, k, file);
+        if (uglifyConfig[dest + '.js']){
+          uglifyConfig[dest + '.js'].push(file);
+        }
+        else {
+          uglifyConfig[dest + '.js'] = [file];
+        }
+        uglifyConfig[path.join('dist','brick.js')].push(file);
+      });
 
     });